[refactor] #39 - CustomTextField 구현
