+++ POTI-iOS/POTI-iOS/Application/SceneDelegate.swift
@@ -15,7 +15,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         guard let windowScene = (scene as? UIWindowScene) else { return }
         let window = UIWindow(windowScene: windowScene)
-        let vc = UINavigationController(rootViewController: FormFieldDemoViewController())
+        let vc = UINavigationController(rootViewController: PotiTabBar())
         window.rootViewController = vc
         self.window = window
         window.makeKeyAndVisible()
